AGENTS address review: generalize skill wording, merge sync helper, expand run.md caching/config docs
