+++ benchopt/skills/using-benchopt/assets/dataset.py
@@ -9,6 +9,8 @@ class Dataset(BaseDataset):
         "n_samples": [100, 1000],
         "n_features": [20],
     }
+    # Small, fast config used by `benchopt test`.
+    test_parameters = {"n_samples": [10], "n_features": [5]}
     # Params that don't affect prepare() output (omit or set to "all"):
     # prepare_cache_ignore = ("seed",)
 
@@ -25,17 +27,3 @@ class Dataset(BaseDataset):
     # or share an _ensure_prepared() guard called from both.
     # def prepare(self):
     #     ...  # idempotent; safe to re-run
-
-
-# Minimal zero-dependency smoke-test dataset — keeps `benchopt test` fast.
-class SimulatedSmall(BaseDataset):
-    name = "simulated-small"
-    requirements = []
-    parameters = {"n_samples": [50], "n_features": [10]}
-    test_parameters = {"n_samples": [10], "n_features": [5]}
-
-    def get_data(self):
-        rng = np.random.default_rng(self.get_seed(use_repetition=True))
-        X = rng.standard_normal((self.n_samples, self.n_features))
-        y = X @ rng.standard_normal(self.n_features)
-        return dict(X=X, y=y)
